feat: additional resource support and improved editing
Pushing additional support for the rest of the Gateway API components as well as mapping to
the backing services and ultimately deployments.

Also updated the editor to handle unrecognized/unsupported YAML inputs and render a collapsible
error block at the top with more information on the YAML issues present, if any.