Expanded Gateway API support and feature improvements (#4)
* feat: additional resource support and improved editing

Pushing additional support for the rest of the Gateway API components as well as mapping to
the backing services and ultimately deployments.

Also updated the editor to handle unrecognized/unsupported YAML inputs and render a collapsible
error block at the top with more information on the YAML issues present, if any.

* chore(deps): updating project deps

* feat: secret detection in the editor block, preventing possible secret leaks

* feat: updating the route coverage table for columnar sorting and pagination

* refactor: code review suggestions

* fix: auto-clearing the secret warning when new input is entered

* fixing banner clearing for secrets and removing some false positive secret-esque values

* adding resource references for secret detection and improving the details pane

* adding version and commit hash into the header